requirements.txt
Flask==1.1.1
google-cloud-datastore==1.7.3
google-cloud-firestore==1.4.0